Revert "Adjust to GDExtension API: functions moving data *into* Godot now accept String/StringName pointers instead of char*"
Didn't work out as expected, because String and StringName are implemented as an opaque byte buffer.
Returning a string_sys() pointer -- which is the *address* of that buffer -- from a callback means that the buffer may not exist anymore inside Godot. Thus, Godot won't be able to read the returned GDNative{String|StringName}Ptr.

This reverts commit 3d5829b22ec5db99294ddb27d3f4005bc3421f54.
This reverts commit a05701b25524e68c562352add0eb2c3b6fde76ca.