Fix 'has_return_value' wrongly inferred by macro -> enables ptrcall!
A `$var:ty` macro variable becomes an atomic unit and can no longer be tokenized or matched against concrete tokens, such as `()`. This is however exactly what `gdext_wrap_method_has_return_value!` was attempting to do, and since the argument was a `ty`, it always resolved to its second branch.