CI: Godot integration tests abort when there's an error
This took far more time to implement than it should have, due to small discrepancies of shells and Godot not behaving 100% as expected with stdout/stderr.
Uses "SCRIPT ERROR:" and not "ERROR:" for now, because the engine occasionally outputs "ERRORs" due to bugs, and this would grind CI to a halt every time.