CI: license-guard uses 'check' instead of 'fix'; remove unneeded checkout@v3 calls
The license-guard mode 'fix' was intended to directly modify pull requests, but this doesn't work for authors that aren't maintainers.
Because there is usually just a small number of files affected, we can simply print the error message.

The checkout@v3 GitHub Action cleans the entire cloned-into directory, and takes longer.
Since the checkout is already done by the parent workflow, it does not need to be repeated.