Support for default parameters, using builder pattern
Assume a function with 3 parameters, 2 of which have default values:

   fn do_sth(num: int, string: GodotString, flag: bool);
   //                  ^ default ""         ^ default false

So far, it had to be called by providing all arguments, default or not:

   do_sth(1, "".into(), false)

With this change, there are now two overloads provided, which allow calling as follows:

   do_sth(1)
   do_sth_ex(1).string("s".into()).flag(true).done()
   do_sth_ex(1).flag(true).done()

The last example shows that it's also possible to skip parameters.
The builder pattern is an emulation of named arguments in Rust.

This implementation requires declaration of a builder struct for each function that has default parameters.

Design choices regarding forward-compatibility (still an improvement over status quo, where everything except rename is breaking.
* New default parameters are added         -> yes
* Existing parameters get a default value  -> no
* A default parameter is renamed           -> no