Add Godot's named colors as constants
All named colors were included using values from Godot's color constant
docs except for `TRANSPARENT`, which is instead `TRANSPARENT_WHITE`.
This is to be more explicit, which here is useful when users are working
with gradients or modulations of color.  This creates a conflict with
`Color::from_string`, which is documented with the colors.

`from_string`'s documentation still references Godot's own color
constants doc since that represents ground truth.

This is executed as a set of hard-coded values because I wasn't keen on
spending the time to learn the code gen system.  This may later be
redone with code gen.

Signed-off-by: Jan Haller <bromeon@gmail.com>