Fix UB for inbound varcalls, if argument count mismatches
It was possible to create undefined behavior by doing dynamic calls to #[func] methods.
If the number of passed arguments was less than the number of declared parameters, the
varcall handler (which delegates to the user-defined #[func]) caused a read past valid
boundaries of the argument array.

This now causes a panic instead. In the future, such errors could be delegated to the
user via Result<T, CallError> return type (maybe under separate `try_*` names).