Provide a default for `Rid`
A default value for `Rid` makes it possible to use it as a member of
`GodotClass`-derived struct, which can be useful for custom Resources.

The default value of `Invalid` is both logical--it shouldn't point to
any resource a user doesn't specify--and it's what's done in-engine
(godot@46dc27/core/templates/rid.h:40 - `uint64_t _id = 0;`).