Support `#[export(range = (radians_as_degrees, suffix="XX"))]`
`radians_as_degrees` was added in Godot [4.2][1] to replace `radians`
with identical functionality and a better name.  `suffix` was added in
[4.0][2] and was seemingly missed when `#[export(range)]` was first
implemented.

This is the first instance where Godot has deprecated an API and gdext
has needed to implement it.  Hopefully this will serve as a useful
template for future work.

This also fixes a previously unknown bug where a `KEY=VALUE` expression
in the third slot of `range = (...)` would yield strange error messages.

[1]: https://github.com/godotengine/godot/pull/82195
[2]: https://github.com/godotengine/godot/pull/50009