Add a doc to point users to kwarg builders
A frequent issue that comes up with people that are first using the
library is that they don't know about gdext's unique model for
Gdscript's default function arguments.  This gently points users to the
builder functions.

Documenting every method this way is a little spammy since users only
need to learn about this API once, but it's only one line, so I don't
think it impacts scrolling distances much on the doc site.