Resolve doc warning with global enums
If you run `cargo doc`, as apposed to `cargo doc -p godot --no-libs` as
in the CI, you get hit with a warning "warning: redundant explicit link
target" on the docs in this commit.  I'm not sure why these options
together supress the warning (neither individually does).  But this
reduces noise in developer consoles.