Remove impl AsObjectArg<T> for Gd<T> values
The idea is to require pass-by-ref for objects, in order to discourage
accidental/unnecessary cloning. This is also consistent with AsArg
by-refs, in particular `impl AsArg<Gd<T>>` when used in arrays and
packed arrays.

The downside is that it needs an extra `&` in some places, and one can
no longer use pass-by-value to "consume" objects, ending their lifetime
with the call. What is also unclear is whether there are situations
where we can benefit from moving "into" the engine (perf-wise). So this
may need some more discussion.

One option is also to move toward using Copy for manually managed Gd<T>
types, which would mean a) those could be passed by-value, and b) we
would anyway lose the "consume" semantics.

This is a more conservative choice for now, with minor syntactic
drawbacks, but in line with AsArg<T> and the potential to expand in the
future.