Discussion about impl AsObjectArg<T> for &Option<Gd<T>> (outer ref)
It's relatively common that Godot APIs return `Option<Gd<T>>` or pass
this type in virtual functions. To avoid excessive `as_ref()` calls, we
**could** directly support `&Option<Gd>` in addition to `Option<&Gd>`.

However, this is currently not done as it hides nullability, especially
in situations where a return type is directly propagated:

    api(create_obj().as_ref())
    api(&create_obj())

While the first is slightly longer, it looks different from a function
create_obj() that returns Gd<T> and thus can never be null. It's also
quite idiomatic to use as_ref() for inner-option transforms in Rust.

For now, the code is still added but inactive.