Remove impl AsObjectArg<T> for &mut Gd<T>
`&mut obj` args can only be passed once, since &mut T: !Copy. This means
that passing it to a Godot API works once, and then causes moved-out
errors. This comes with a few issues:

* It either needs the `&*` pattern or not, depending on the order of
  calls. Reordering breaks them.

* It's inconsistent with `AsArg`, which is only implemented for `&T` in
  case of by-ref builtins.

* Consuming `&mut T` should probably be allowed iff consuming `T` is
  allowed.

That said, we'll need to check the ergonomic impact and see how
arguments are being passed in practice.