Added ConnectHandle for disconnections
Changes some of the `connect_` methods for typed signals to return a handle containing information about the connection.
Can be used to disconnect the signal when not needed. Makes it simpler to disconnect signals to objects that are dead / about to die:
Object can call disconnect on handles before being freed (for example in a "impl Drop" implementation).