GDScript test suite now has tests that fail until confirmed at end
Necessary because some operations cause Godot errors, which cause immediate return from the
`test_*` GDScript function. However, the runner cannot differentiate that from a legitimate
return at the end, as the only difference is a printed error message (which isn't accessible
programmatically).

As a result, a new pattern with mark_test_pending() + mark_test_succeeded() is introduced.
If the former method is called without the latter, the test is considered failed.