Derive `Hash` for `Rect2i`
Also add `Hash`, `PartialOrd` and `Ord` for some enums while I have my
eyes on them.

Related issue: #505. There seems to be no objective to keep Godot and
Rust hashes the same, so this change should be fine, and is indeed
consistent with prior art of deriving `Hash` on e.g. `Vector2i`.