Add bounds for future-proof `Gd` deref
The current `RawGd::as_upcast_ref/mut` rely on `mem::transmute` to convert between
`&Gd<T>` and `&Base`. This works as long as Base is an engine-defined class, which
is currently always given.

However, if Rust-Rust inheritance is introduced at some point in the future,
this assumption no longer holds. The new bound ensures soundness in such cases.