Add Base::to_init_gd() to hand out early pointers
For manually-managed objects, the implementation is simple: just create a copy
of the internal Gd pointer.

For ref-counted objects, we need to deal with the problem that `Base<T>` stores
a weak Gd pointer, and thus dropping handed-out pointers will trigger object
destruction. To deal with this, we keep an extra strong pointer around, and
defer its destruction to the next frame.

It would be even better if we could immediately destroy the strong ref after
object construction, however we could only do so for Rust paths `new_gd()`,
`Gd::from_init_fn()`, not for Godot-side construction. Unfortunately, Godot
doesn't offer a hook to run code immediately after creation. The POSTINITIALIZE
notification needs to be dispatched by godot-rust, and thus suffers from the
same issue.