Fix documentation of `InitLevel`
We previously used a type alias because `use` statements don't allow RustDoc in this case.
However, this caused many broken links and shows up in a "Type Aliases" section, rather than
in "Enums" next to `InitStage`.

The new setup requires HTML links because the godot-ffi crate doesn't know the symbols, but
it ultimately works, and we can also avoid duplicating docs.