Remove `ToGodot`/`FromGodot` impls for `u64`
Disallows `u64` from being used in `#[func]`. While this was allowed so far, it is very error-prone, since
GDScript cannot represent 64-bit unsigned integers and may silently convert to a different value. Godot's
`Variant` type cannot store the upper half of values of `u64`, either.

So far, `ToGodot::to_godot()` caused a panic, but in doing so it deviates from the contract and expectations.
It is safer to use `i64` directly, or write a newtype with manual `ToGodot/FromGodot` impls and handling
out-of-range values explicitly.

Since `u64` is still used in engine APIs and bitfields -- where it is completely safe, since the backing code is
C++ with `uint64_t` -- the codegen/macro code needs special-casing. Not ideal, but the alternative of having yet
another `ToGodot/FromGodot` trait with just this exception isn't great, either. Maybe over time we can revisit this.